requirements.txt
Pillow>=10.0.0
pillow-heif>=0.13.0


